Moving classes out of the old packages.
