+++ parquet-column/src/main/java/org/apache/parquet/column/page/PageWriter.java
@@ -20,7 +20,7 @@ package org.apache.parquet.column.page;
 
 import java.io.IOException;
 
-import parquet.bytes.ByteBufferAllocator;
+import org.apache.parquet.bytes.ByteBufferAllocator;
 import org.apache.parquet.bytes.BytesInput;
 import org.apache.parquet.column.Encoding;
 import org.apache.parquet.column.statistics.Statistics;
