+++ parquet-column/src/main/java/org/apache/parquet/column/values/boundedint/BitWriter.java
@@ -19,7 +19,7 @@
 package org.apache.parquet.column.values.boundedint;
 
 import org.apache.parquet.ParquetRuntimeException;
-import parquet.bytes.ByteBufferAllocator;
+import org.apache.parquet.bytes.ByteBufferAllocator;
 import org.apache.parquet.Log;
 import org.apache.parquet.bytes.BytesInput;
 import org.apache.parquet.bytes.CapacityByteArrayOutputStream;
