+++ parquet-column/src/main/java/org/apache/parquet/column/values/plain/PlainValuesReader.java
@@ -23,7 +23,7 @@ import static org.apache.parquet.Log.DEBUG;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 
-import parquet.bytes.ByteBufferInputStream;
+import org.apache.parquet.bytes.ByteBufferInputStream;
 import org.apache.parquet.Log;
 import org.apache.parquet.bytes.LittleEndianDataInputStream;
 import org.apache.parquet.column.values.ValuesReader;
