+++ parquet-column/src/test/java/org/apache/parquet/column/values/deltalengthbytearray/TestDeltaLengthByteArray.java
@@ -19,13 +19,11 @@
 package org.apache.parquet.column.values.deltalengthbytearray;
 
 import java.io.IOException;
-import java.util.Arrays;
 
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.Assert;
 
-import parquet.bytes.DirectByteBufferAllocator;
+import org.apache.parquet.bytes.DirectByteBufferAllocator;
 import org.apache.parquet.column.values.Utils;
 import org.apache.parquet.column.values.ValuesReader;
 import org.apache.parquet.column.values.delta.DeltaBinaryPackingValuesReader;
