+++ parquet-common/src/main/java/org/apache/parquet/bytes/ByteBufferInputStream.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package parquet.bytes;
+package org.apache.parquet.bytes;
 
 import java.io.IOException;
 import java.io.InputStream;
