+++ parquet-encoding/src/main/java/org/apache/parquet/bytes/CapacityByteArrayOutputStream.java
@@ -21,7 +21,6 @@ package org.apache.parquet.bytes;
 import static java.lang.Math.max;
 import static java.lang.Math.pow;
 import static java.lang.String.format;
-import static java.lang.System.arraycopy;
 import static org.apache.parquet.Preconditions.checkArgument;
 
 import java.io.ByteArrayOutputStream;
@@ -32,7 +31,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.parquet.Log;
-import parquet.bytes.ByteBufferAllocator;
 
 /**
  * Similar to a {@link ByteArrayOutputStream}, but uses a different strategy for growing that does not involve copying.
